Move property tests over to tests/unit
Based on review feedback.  This also makes it easier to add
property tests as functional and integration tests.

I've set the default number of iterations to 10 by default.
On travis, we bump that up to 2000 via the HYPOTHESIS_PROFILE
env var.
